# secp256k1: Implement direct signature verification.