**This requires PR #2057**.

This implements signature verification directly instead of calling the version in the standard library in order to ultimately be able to break the reliance on big ints it requires and thus allow for optimization specifically for the sec256k1 curve.

It currently is only marginally faster since it still relies on big ints.  However, there is already a 10% reduction in allocations which reduces pressure on the GC during IBD when there are a ton of signature verifications happening.

```
benchmark            old ns/op    new ns/op    delta
-----------------------------------------------------
BenchmarkSigVerify   224238       219603       -2.07%

benchmark            old allocs   new allocs   delta
-----------------------------------------------------
BenchmarkSigVerify   64           57           -10.94%

benchmark            old bytes    new bytes    delta
-----------------------------------------------------
BenchmarkSigVerify   3138         2818         -10.20%
```